# docs/openapi/teamclaw-api.v1.yaml
openapi: 3.1.0
info:
  title: TeamClaw Cloud API
  version: 1.0.0
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: https://api.teamclaw.local
security:
  - bearerAuth: []
...
paths:
...
  /v1/teams/{teamId}/actors:
    get:
      operationId: listTeamActors
      summary: List team actors
      tags: [Actors]
      parameters:
        - $ref: "#/components/parameters/TeamId"
        - name: kind
          in: query
          schema:
            type: string
            enum: [agent, user, external]
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 500
            default: 200
      responses:
        "200":
          description: Actor page.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActorPage"
        "401":
          $ref: "#/components/responses/Error"
  /v1/actors/{actorId}:
    get:
      operationId: getActor
      summary: Get single actor
      tags: [Actors]
      parameters:
        - name: actorId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Actor.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Actor"
        "404":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
  /v1/actors/external:
    post:
      operationId: upsertExternalActor
      summary: Upsert external actor
      tags: [Actors]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [teamId, source, sourceId, displayName]
              properties:
                teamId:
                  type: string
                  format: uuid
                source:
                  type: string
                sourceId:
                  type: string
                displayName:
                  type: string
      responses:
        "200":
          description: Upserted actor.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                type: object
                required: [actorId]
                properties:
                  actorId:
                    type: string
                    format: uuid
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
  /v1/teams/{teamId}/agents/connected:
    get:
      operationId: listConnectedAgents
      summary: List connected agents
      tags: [Actors]
      parameters:
        - $ref: "#/components/parameters/TeamId"
      responses:
        "200":
          description: Connected agents.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/ConnectedAgent"
        "401":
          $ref: "#/components/responses/Error"
  /v1/agents/{agentActorId}:
    patch:
      operationId: updateOwnedAgentProfile
      summary: Update owned agent profile
      tags: [Actors]
      parameters:
        - name: agentActorId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentProfilePatch"
      responses:
        "204":
          description: OK
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
  /v1/agents/{agentActorId}/defaults:
    patch:
      operationId: updateAgentDefaults
      summary: Update agent defaults
      tags: [Actors]
      parameters:
        - name: agentActorId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentDefaultsPatch"
      responses:
        "204":
          description: OK
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
  /v1/agents/{agentActorId}/permission:
    get:
      operationId: checkAgentPermission
      summary: Check agent permission
      tags: [Actors]
      parameters:
        - name: agentActorId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: actorId
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Permission check result.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                type: object
                required: [allowed]
                properties:
                  allowed:
                    type: boolean
                  role:
                    type:
                      - string
                      - "null"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
  /v1/agents/{agentActorId}/access:
    get:
      operationId: listAgentAccess
      summary: List agent access
      tags: [Actors]
      parameters:
        - name: agentActorId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Agent access list.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/AgentAccess"
        "401":
          $ref: "#/components/responses/Error"
    post:
      operationId: grantAgentAccess
      summary: Grant agent access
      tags: [Actors]
      parameters:
        - name: agentActorId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [actorId, role]
              properties:
                actorId:
                  type: string
                  format: uuid
                role:
                  type: string
      responses:
        "200":
          description: Granted access.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentAccess"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
  /v1/agents/{agentActorId}/access/{actorId}:
    delete:
      operationId: revokeAgentAccess
      summary: Revoke agent access
      tags: [Actors]
      parameters:
        - name: agentActorId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: actorId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: OK
        "401":
          $ref: "#/components/responses/Error"
  /v1/agents/{agentActorId}/admin-members:
    get:
      operationId: listAgentAdminMembers
      summary: List agent admin member actor IDs
      tags: [Actors]
      parameters:
        - name: agentActorId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Admin member IDs.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      type: string
                      format: uuid
        "401":
          $ref: "#/components/responses/Error"
...
  /v1/ideas:
    get:
      operationId: listIdeas
      summary: List ideas
      tags: [Ideas]
      parameters:
        - in: query
          name: teamId
          required: true
          schema:
            type: string
            format: uuid
        - in: query
          name: archived
          schema:
            type: boolean
            default: false
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
        - in: query
          name: cursor
          schema:
            type: string
      responses:
        "200":
          description: Idea page.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IdeaPage"
        "401":
          $ref: "#/components/responses/Error"
    post:
      operationId: createIdea
      summary: Create idea
      tags: [Ideas]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IdeaCreate"
      responses:
        "201":
          description: Created idea.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Idea"
        "401":
          $ref: "#/components/responses/Error"
  /v1/ideas/{ideaId}:
    get:
      operationId: getIdea
      summary: Get idea
      tags: [Ideas]
      parameters:
        - in: path
          name: ideaId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Idea.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Idea"
        "404":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
    patch:
      operationId: updateIdea
      summary: Update idea
      tags: [Ideas]
      parameters:
        - in: path
          name: ideaId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IdeaUpdate"
      responses:
        "200":
          description: Updated idea.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Idea"
        "401":
          $ref: "#/components/responses/Error"
  /v1/ideas/{ideaId}/archive:
    post:
      operationId: archiveIdea
      summary: Archive idea
      tags: [Ideas]
      parameters:
        - in: path
          name: ideaId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: OK
        "401":
          $ref: "#/components/responses/Error"
  /v1/ideas/{ideaId}/activities:
    post:
      operationId: createIdeaActivity
      summary: Create idea activity
      tags: [Ideas]
      parameters:
        - in: path
          name: ideaId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [kind, actorId]
              properties:
                kind:
                  type: string
                content:
                  type:
                    - string
                    - "null"
                actorId:
                  type: string
                  format: uuid
                metadata:
                  type:
                    - object
                    - "null"
                  additionalProperties: true
      responses:
        "201":
          description: Created activity.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IdeaActivity"
        "401":
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  headers:
    RequestId:
      schema:
        type: string
  parameters:
    TeamId:
      name: teamId
      in: path
      required: true
      schema:
        type: string
...
  responses:
...
    Error:
      description: Error response.
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties:
            code:
              type: string
              enum:
                - missing_auth
                - invalid_json
                - validation_failed
                - forbidden
                - not_found
                - conflict
                - rate_limited
                - upstream_unavailable
                - internal
            message:
              type: string
            requestId:
              type: string
...
    Actor:
      type: object
      required: [id, teamId, kind, displayName]
      properties:
        id:
          type: string
          format: uuid
        teamId:
          type: string
          format: uuid
        kind:
          type: string
          enum: [agent, user, external]
        displayName:
          type: string
        avatarUrl:
          type:
            - string
            - "null"
        metadata:
          type:
            - object
            - "null"
          additionalProperties: true
    ActorPage:
      type: object
      required: [items, nextCursor]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Actor"
        nextCursor:
          type:
            - string
            - "null"
    ConnectedAgent:
      allOf:
        - $ref: "#/components/schemas/Actor"
        - type: object
          properties:
            deviceId:
              type:
                - string
                - "null"
            lastSeenAt:
              type:
                - string
                - "null"
              format: date-time
            agentType:
              type: string
    AgentProfilePatch:
      type: object
      properties:
        displayName:
          type: string
        avatarUrl:
          type:
            - string
            - "null"
        description:
          type:
            - string
            - "null"
    AgentDefaultsPatch:
      type: object
      properties:
        defaultAgentType:
          type: string
        supportedAgentTypes:
          type: array
          items:
            type: string
    AgentAccess:
      type: object
      required: [agentActorId, actorId, role]
      properties:
        agentActorId:
          type: string
          format: uuid
        actorId:
          type: string
          format: uuid
        role:
          type: string
...
    Idea:
      type: object
      required: [id, teamId, title, archived, createdAt]
      properties:
        id:
          type: string
          format: uuid
        teamId:
          type: string
          format: uuid
        title:
          type: string
        description:
          type:
            - string
            - "null"
        archived:
          type: boolean
        authorActorId:
          type:
            - string
            - "null"
          format: uuid
        actorIds:
          type: array
          items:
            type: string
            format: uuid
        createdAt:
          type:
            - string
            - "null"
          format: date-time
        updatedAt:
          type:
            - string
            - "null"
          format: date-time
...
    IdeaCreate:
      type: object
      required: [teamId, title, authorActorId]
      properties:
        id:
          type: string
          format: uuid
        teamId:
          type: string
          format: uuid
        title:
          type: string
        description:
          type:
            - string
            - "null"
        authorActorId:
          type: string
          format: uuid
        actorIds:
          type: array
          items:
            type: string
            format: uuid
    IdeaUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type:
            - string
            - "null"
        actorIds:
          type: array
          items:
            type: string
            format: uuid
    IdeaPage:
      type: object
      required: [items, nextCursor]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Idea"
        nextCursor:
          type:
            - string
            - "null"
    IdeaActivity:
      type: object
      required: [id, ideaId, kind, actorId, createdAt]
      properties:
        id:
          type: string
          format: uuid
        ideaId:
          type: string
          format: uuid
        kind:
          type: string
        content:
          type:
            - string
            - "null"
        actorId:
          type: string
          format: uuid
        metadata:
          type:
            - object
            - "null"
          additionalProperties: true
        createdAt:
          type: string
          format: date-time
...