# services/fc/lib/routes/ideas.mjs
import { ApiError } from "../http-utils.mjs";
import { requireString, encodeCursor } from "../router.mjs";

const DEFAULT_IDEA_LIMIT = 50;
const MAX_IDEA_LIMIT = 200;

function parseIdeaLimit(value) {
  if (value === null || value === undefined || value === "") return DEFAULT_IDEA_LIMIT;
  const limit = Number(value);
  if (!Number.isInteger(limit) || limit < 1 || limit > MAX_IDEA_LIMIT) {
    throw new ApiError(400, "validation_failed", "limit must be an integer from 1 to 200");
  }
  return limit;
}

function decodeIdeaCursor(value) {
  if (!value) return null;
  try {
    const parsed = JSON.parse(Buffer.from(value, "base64url").toString("utf8"));
    if (!parsed || typeof parsed !== "object") throw new Error("not an object");
    return {
      updatedAt: parsed.updatedAt ?? null,
      id: parsed.id ?? null,
    };
  } catch (cause) {
    throw new ApiError(400, "validation_failed", "Invalid cursor", { cause });
  }
}

export function nextIdeaCursor(items, limit) {
  if (!Array.isArray(items) || items.length < limit) return null;
  const last = items[items.length - 1];
  if (!last) return null;
  return encodeCursor({
    updatedAt: last.updatedAt ?? null,
    id: last.id,
  });
}

export function registerIdeas(router) {
  router.get("/v1/ideas", async (ctx) => {
    const teamId = ctx.query.get("teamId");
    if (!teamId) throw new ApiError(400, "validation_failed", "teamId is required");
    const limit = parseIdeaLimit(ctx.query.get("limit"));
    const archived = ctx.query.get("archived") === "true";
    const cursorStr = ctx.query.get("cursor");
    const cursor = cursorStr ? decodeIdeaCursor(cursorStr) : null;
    const page = await ctx.repository.listIdeas({ teamId, archived, limit, cursor });
    const nextCursor = nextIdeaCursor(page.items, limit);
    return { body: { items: page.items, nextCursor } };
  });

  router.post("/v1/ideas", async (ctx) => {
    const body = ctx.json ?? {};
    requireString(body.teamId, "teamId");
    requireString(body.title, "title");
    requireString(body.authorActorId, "authorActorId");
    const idea = await ctx.repository.createIdea(body);
    return { statusCode: 201, body: idea };
  });

  router.get("/v1/ideas/:ideaId", async (ctx) => {
    const idea = await ctx.repository.getIdea(ctx.params.ideaId);
    if (!idea) throw new ApiError(404, "not_found", "idea not found");
    return { body: idea };
  });

  router.patch("/v1/ideas/:ideaId", async (ctx) => {
    const idea = await ctx.repository.updateIdea(ctx.params.ideaId, ctx.json ?? {});
    if (!idea) throw new ApiError(404, "not_found", "idea not found");
    return { body: idea };
  });

  router.post("/v1/ideas/:ideaId/archive", async (ctx) => {
    await ctx.repository.archiveIdea(ctx.params.ideaId);
    return { statusCode: 204, body: null };
  });

  router.post("/v1/ideas/:ideaId/activities", async (ctx) => {
    const body = ctx.json ?? {};
    requireString(body.kind, "kind");
    requireString(body.actorId, "actorId");
    const activity = await ctx.repository.createIdeaActivity(ctx.params.ideaId, body);
    return { statusCode: 201, body: activity };
  });
}